+++ src/config/index.js
@@ -31,6 +31,7 @@ import { loadables } from './Loadables';
 
 import { sentryOptions } from './Sentry';
 import { contentIcons } from './ContentIcons';
+import { toolbar } from './toolbar';
 
 import applyAddonConfiguration from 'load-volto-addons';
 
@@ -118,6 +119,7 @@ let config = {
 
   addonRoutes: [],
   addonReducers: {},
+  toolbar,
 };
 
 config = applyAddonConfiguration(config);
@@ -129,3 +131,4 @@ export const blocks = config.blocks;
 export const addonRoutes = [...config.addonRoutes];
 export const addonReducers = { ...config.addonReducers };
 export const appExtras = config.appExtras;
+export const toolbar = { ...config.toolbar };
